# api-filters/src/swaggerDef.json
{
  "swagger": "2.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Emssions API",
    "description": "Get emissions info",
    "license": {
      "name": "GIS",
      "url": "http://girlsincode.com/"
    }
  },
  "basePath": "/",
  "host": "localhost:4000",
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
  "tags": [
    {
      "name": "Emissions",
      "description": "Get emissions"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1": {
      "get": {
        "tags": [
          "emissions"
        ],
        "summary": "get all emissions data",
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/api/v1/countries/{name}": {
      "parameters": [
        {
          "name": "name",
          "in": "path",
          "required": true,
          "description": "country code name (3letters)",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "emissions"
        ],
        "summary": "sum all years emmissions for a country",
        "responses": {
          "200": {
            "description": "ok"
          },
          "400": {
            "description": "failed. no data"
          }
        }
      }
    },
    "/countries": {
      "get": {
        "tags": [
          "emissions"
        ],
        "description": "get all countries sorted asc",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful get response"
          }
        }
      }
    }
  },
  "definitions": {
    "Emission": {
      "type": "object",
      "properties": {
        "year": {
          "type": "number"
        },
        "value": {
          "type": "number"
        }
      }
    },
    "Emissions": {
      "type": "object",
      "properties": {
        "country": {
          "type": "string"
        },
        "sector": {
          "type": "string"
        },
        "parent_sector": {
          "type": "string"
        },
        "emissions": {
          "type": "array"
        }
      }
    }
  },
  "responses": {},
  "parameters": {},
  "securityDefinitions": {}
}